refactor: App layout and onboarding logic
